emugl: Rename GLES dispatch headers and sources.
The headers used to hold the dispatch tables for GLESv1 and GLESv2 are
wrongly called "GLDispatch.h" and "GL2Dispatch.h" respectively. This
patch changes their names to "GLESv1Dispatch.h" and "GLESv2Dispatch.h"
for clarity.

Note that we don't use "GLEScm" for GLES 1.x, because the "CM" stands
for "Common Profile" and is a relic from the initial versions of the API
which we don't really care about. "GLESv1" is just clearer for everyone.

NOTE: This does *not* change the name of the global variables, these
      will be changed in a future patch.

Change-Id: I0a0b687de56616e64b3937aa69b9185085891c9c