emugl: Remove un-needed EGL entry points from libOpenglRender.
libOpenglRender only uses a small subset of the EGL API, so only list
the functions that are actually used by it in render_egl*.entries files.

The main reason to do this is to reduce the cost of implementing an EGL
wrapper on top of ANGLE's EGL library (i.e. less functions to wrap,
implement and maintain).

A future patch will also remove the corresponding code from the EGL
translation library to (slightly) simplify it.

Change-Id: If5acf42b7fad9243f87ef4ea0167d717c6bb1bab