GL encoder null pointer check before write
A related patch is also in the system image.

Previously when encoding some functions, if the user gives us a null
pointer, the encoder will write a header with some data length while not
actually writing any data. This patch fix it by writing 0 into the data
length field for all those functions.

Change-Id: I8b87012d5bcb59fa67fcc8f65a3ca260a102afd2