Checksum for GLES messages from host to guest
This patch adds checksum for GLES messages from host to guest. It also
changes ChecksumCalculator a little because of how encoder and decoder
handle data buffers differently.

Currently when using ChecksumCalculator, for both read and write
checksum, one needs to call addBuffer for each data buffer, then call
validate/write. The reason is the encoder usually has data in multiple
buffers for both read and write, so we can't use the previous validate
interface where it only calculates checksum from one single buffer.

The interface for the decoder, ChecksumCalculatorThreadInfo, currently
provides functions to validate/write checksum for one single buffer.
This is because the decoder always has a large buffer to contain all
data for each GL command for both read and write.

Change-Id: I4a32e33f22b969e522c7bee8fd2c25c674ba1b8b