distrib/android-emugl: Move Android GPU emulation sources here.
This patch is used to add the unmodified sources of the GPU
emulation libraries that are normally under sdk/emulator/opengl/.

These sources were originally built with the platform build system,
but this is no longer the case. Since they are only built through
the emulator's own build system (i.e. android-rebuild.sh), it's
easier to put them under external/qemu/ to ensure that they are
always in sync with the rest of the emulator.

Note that the encoder sources are still in the platform under
device/generic/goldfish/opengl/.

A future patch will update the emulator build system to pick
the sources from here, and another one will remove the sources
from sdk/ after that.

Note: these are the unmodified sources from the following
      commit in https://android.googlesource.com/platform/sdk.git :

    182e469 emulator/opengl: Fix Windows cross-build.

Change-Id: If2010577ef4af4c6755ab345f424af0b546f9282